consolidate tablesorts and put it in ui/lib
