+++ ui/analyse/package.json
@@ -19,8 +19,7 @@
     "lib": "workspace:*",
     "prop-types": "^15.8.1",
     "shepherd.js": "^11.2.0",
-    "sortablejs": "^1.15.6",
-    "tablesort": "^5.5.0"
+    "sortablejs": "^1.15.6"
   },
   "build": {
     "bundle": "src/**/analyse.*ts",
