+++ ui/bits/package.json
@@ -37,7 +37,6 @@
     "prop-types": "^15.8.1",
     "qrcode": "^1.5.4",
     "shepherd.js": "^11.2.0",
-    "tablesort": "^5.3.0",
     "zxcvbn": "^4.4.2"
   },
   "build": {
