+++ ui/bits/src/bits.clas.ts
@@ -1,5 +1,4 @@
-import tablesort from 'tablesort';
-import extendTablesortNumber from 'lib/tablesortNumber';
+import { sortTable, extendTablesortNumber } from 'lib/tablesort';
 import * as xhr from 'lib/xhr';
 import { Textcomplete } from '@textcomplete/core';
 import { TextareaEditor } from '@textcomplete/textarea';
@@ -8,7 +7,7 @@ import type { UserCompleteResult } from 'lib/userComplete';
 
 site.load.then(() => {
   $('table.sortable').each(function (this: HTMLElement) {
-    tablesort(this, {
+    sortTable(this, {
       descending: true,
     });
   });
