+++ ui/lib/package.json
@@ -28,7 +28,7 @@
     "stockfish-nnue.wasm": "1.0.0-1946a675.smolnet",
     "stockfish.js": "^10.0.2",
     "stockfish.wasm": "^0.10.0",
-    "tablesort": "^5.3.0"
+    "tablesort": "5.5.0"
   },
   "build": {
     "sync": {
