+++ ui/mod/src/mod.games.ts
@@ -1,5 +1,4 @@
-import extendTablesortNumber from 'lib/tablesortNumber';
-import tablesort from 'tablesort';
+import { sortTable, extendTablesortNumber } from 'lib/tablesort';
 import { debounce } from 'lib/async';
 import { formToXhr } from 'lib/xhr';
 import { checkBoxAll, expandCheckboxZone, shiftClickCheckboxRange } from './checkBoxes';
@@ -26,7 +25,7 @@ const setupFilter = () => {
 const setupTable = () => {
   const table = document.querySelector('table.game-list') as HTMLTableElement;
   extendTablesortNumber();
-  tablesort(table, { descending: true });
+  sortTable(table, { descending: true });
 
   expandCheckboxZone(table, 'td:first-child', shiftClickCheckboxRange(table));
   checkBoxAll(table);
