+++ ui/site/package.json
@@ -8,8 +8,7 @@
   "dependencies": {
     "lib": "workspace:*",
     "dialog-polyfill": "0.5.6",
-    "lichess-pgn-viewer": "^2.4.0",
-    "tablesort": "^5.3.0"
+    "lichess-pgn-viewer": "^2.4.0"
   },
   "build": {
     "bundle": [
